test: remove mock
