+++ package.json
@@ -33,7 +33,7 @@
     "lint": "biome lint --write",
     "format": "biome format --write",
     "lint:fix": "biome check --write",
-    "test:unit": "bun test test/unit --preload ./happydom.ts",
+    "test:unit": "bun test test/unit --preload ./happydom.ts --isolate --randomize",
     "test:e2e": "playwright test",
     "test:report": "playwright show-report",
     "generate": "drizzle-kit generate",
